requirements.txt
python-dotenv
google-generativeai
pillow
pyautogui
opencv-python>=4.5.5
numpy>=1.22.0